# Speed up Home initialization