--- a/src/containers/ReloadFocus.jsx
+++ b/src/containers/ReloadFocus.jsx
@@ -1,7 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-
-import { withClient, Q } from 'cozy-client'
+import {
+  makeTriggersWithJobStatusQuery,
+  makeAppsQuery,
+  makeJobsQuery
+} from 'queries'
+import { withClient } from 'cozy-client'
 class RealoadFocus extends React.Component {
   static contextTypes = {
     store: PropTypes.object
@@ -9,13 +13,12 @@ class RealoadFocus extends React.Component {
   componentDidMount() {
     const client = this.props.client
     window.addEventListener('focus', () => {
-      client.query(Q('io.cozy.jobs'))
+      client.query(makeJobsQuery.definition(), makeJobsQuery.options)
       client.query(
-        Q('io.cozy.triggers').where({
-          worker: { $in: ['client', 'konnector'] }
-        })
+        makeTriggersWithJobStatusQuery.definition(),
+        makeTriggersWithJobStatusQuery.options
       )
-      client.query(Q('io.cozy.apps'))
+      client.query(makeAppsQuery.definition(), makeAppsQuery.options)
     })
   }