… package.json
… src/dataproxy/DataProxyProvider.jsx
… yarn.lock