This is meant to speed up to Home initialization and thus its display, particularly for the flagship app.

Measured time: 
[Web]
- Before : ~576ms
- After: ~130ms

[React-Native]
- Before : ~719ms
- After: ~178ms 

For both environments, this is roughly a 4x in performances.

```
### ✨ Features

* Improve starting time

### 🐛 Bug Fixes

*

### 🔧 Tech

*
```